Added my tests into copy of teammates dominion folder
