+++ projects/judges/edwardseDominion/asserttrue.h
@@ -1,0 +1,11 @@
+#include <stdlib.h>
+
+void assertTrue(int test){
+	if(test) {
+    	printf("Test was Successful\n\n");
+    } else {
+    	printf("Test Failed.\n\n");
+    }
+
+    return;
+}
